Issue 221, make BatchProvider Worker.
